backend/requirements.txt
Django==4.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
requests==2.28.1
django-cors-headers==3.13.0
reportlab==4.0.4
cryptography==41.0.4
endesive==2.0.12
pyopenssl==23.2.0
certifi==2023.7.22
